harvester.py broken down into modules in the motesio package
